pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
Flask = "^2.1.3"
openai = "^0.27.9"
bcrypt = "^4.0.1"
gevent = "^23.7.0"
Flask-Compress = "^1.13"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}
toml = "^0.10.2"
poetry = {url = "https://storage.googleapis.com/poetry-bundles/poetry-1.1.15-py2.py3-none-any.whl"}
urllib3 = "1.26.15"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"